+++ pkg/tasks/kubernetes_binaries.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package upgrade
+package tasks
 
 import (
 	"github.com/pkg/errors"
