+++ pkg/tasks/util.go
@@ -14,10 +14,11 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package upgrade
+package tasks
 
 import (
 	"context"
+	"time"
 
 	"github.com/pkg/errors"
 
@@ -31,6 +32,17 @@ import (
 	"k8s.io/client-go/util/retry"
 	dynclient "sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
+)
+
+const (
+	labelUpgradeLock      = "kubeone.io/upgrade-in-progress"
+	labelControlPlaneNode = "node-role.kubernetes.io/master"
+	// timeoutKubeletUpgrade is time for how long kubeone will wait after upgrading kubelet
+	// and running the upgrade process on the node
+	timeoutKubeletUpgrade = 1 * time.Minute
+	// timeoutNodeUpgrade is time for how long kubeone will wait after finishing the upgrade
+	// process on the node
+	timeoutNodeUpgrade = 30 * time.Second
 )
 
 func determineHostname(s *state.State) error {
@@ -53,7 +65,7 @@ func determineHostname(s *state.State) error {
 
 		node.SetHostname(stdout)
 		return nil
-	}, true)
+	}, state.RunParallel)
 }
 
 func determineOS(s *state.State) error {
@@ -66,7 +78,7 @@ func determineOS(s *state.State) error {
 
 		node.SetOperatingSystem(osID)
 		return nil
-	}, true)
+	}, state.RunParallel)
 }
 
 func labelNode(client dynclient.Client, host *kubeoneapi.HostConfig) error {
